requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai>=1.12.0
supabase==1.0.3
httpx>=0.23.0,<0.24.0
pydantic==2.6.1
pydantic-settings==2.2.1
python-multipart==0.0.9
langdetect==1.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 